backend/requirements.txt
django-ninja
django-cors-headers
langchain
python-dotenv
jq
openai
qdrant-client
django_huey
redis
tiktoken